requirements.txt
amqp==5.2.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
billiard==4.2.0
blinker==1.7.0
celery==5.3.6
cffi==1.16.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
Flask==3.0.2
Flask-SQLAlchemy==3.1.1
greenlet==3.0.3
itsdangerous==2.1.2
Jinja2==3.1.3
kombu==5.3.5
MarkupSafe==2.1.5
packaging==23.2
passlib==1.7.4
prompt-toolkit==3.0.43
pycparser==2.21
PyJWT==2.8.0
python-dateutil==2.9.0.post0
redis==5.0.3
setuptools==69.0.3
six==1.16.0
SQLAlchemy==2.0.27
typing_extensions==4.9.0
tzdata==2024.1
vine==5.1.0
wcwidth==0.2.13
Werkzeug==3.0.1
wheel==0.42.0
